requirements.txt
requests 
python-dotenv